refactor: modify dark mode media query hooks directory
